LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-buttons-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>IOSInlineMediaControls</code> computed styles.");

window.jsTestIsAsync = true;

const mediaControls = new IOSInlineMediaControls({ width: 680, height: 300 });
document.body.appendChild(mediaControls.element);

const buttonPositions = [
    [mediaControls.playPauseButton, 12],
    [mediaControls.skipBackButton, 10],
    [mediaControls.airplayButton, 13],
    [mediaControls.pipButton, 13],
    [mediaControls.fullscreenButton, 13]
];

let button, y;
(function testButton(index)
{
    [button, y] = buttonPositions[index];
    debug("");
    debug(button.constructor.name);
    shouldBecomeEqualToString("getComputedStyle(button.element).webkitMaskPositionY", `${y}px`, () => {
        shouldBeEqualToRGBAColor("getComputedStyle(button.element).backgroundColor", "rgb(0, 0, 0)");
        shouldBeEqualToString("getComputedStyle(button.element).height", "50px");
        if (index === buttonPositions.length - 1) {
            mediaControls.element.remove();
            finishJSTest();
        } else
            testButton(index + 1);
    });
})(0);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
